pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "surya-ocr-pro"
version = "3.0.0"
description = "High-precision document OCR powered by Surya/marker-pdf"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "marker-pdf>=1.10",
    "PyMuPDF>=1.24",
    "Pillow>=10.0",
    "python-docx>=1.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
gui = ["customtkinter>=5.2"]
dev = ["pytest>=8.0", "pytest-cov", "ruff"]

[project.scripts]
surya-ocr = "surya_ocr.cli:main"

[project.gui-scripts]
surya-ocr-gui = "surya_ocr.gui.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/surya_ocr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that require the model (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 120
target-version = "py310"
